Move client classes to a new package
