+++ backend/src/main/java/pt/ua/deti/ies/SmartHomes/backend/Clients/ClientRepository.java
@@ -1,4 +1,4 @@
-package pt.ua.deti.ies.SmartHomes.backend.Database;
+package pt.ua.deti.ies.SmartHomes.backend.Clients;
 
 import org.springframework.data.jpa.repository.JpaRepository;
