+++ backend/src/main/java/pt/ua/deti/ies/SmartHomes/backend/Clients/ClientService.java
@@ -1,4 +1,4 @@
-package pt.ua.deti.ies.SmartHomes.backend.Database;
+package pt.ua.deti.ies.SmartHomes.backend.Clients;
 
 import java.util.List;
